kernel.sh: add _F_scm_git_cloneopts
Usage is easy: if you regularly build a kernel from git, then it's handy
to have a clone under ~/git, then you don't have to re-download the
kernel repo again and again, as long as you're doing a host build.

The option is not set if the dir under ~/git is missing.

NOTE: You are not forced to keep the clone in ~/git up to date, having
an outdated repo can speed up the build as well.