fonts schema: add a new mode for when _F_fonts_subdir is not defined.
if this behavior is encountered, it triggers a new mode that determines
the correct font directory based on the file extension. this allows
proper support of font packs with more than one font type in the packaging.
install scriptlet also now handles the directories as an "array", so it
will work with 1 or more directories instead of just 1. old usage should
still work.